fix(oracle): only get ip address from primary vnic
Previously, the first vnic returned by the oci sdk is used to get the
instance public IP from, but if a secondary vnic is created and
attached after instance creation, this causes this logic to fail
since the new secondary vnic is returned before the primary vnic
by the oci sdk. This commit makes the ip property check for the
is_primary property on all vnics before getting an IP address from
the selected vnic.